+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### November 25, 2022 (DESCW-708) Projects Backend Reorg
 - Backend
   - Reorganized backend for projects (Run npm i to install new dependencies)
+  - Reorganized backend for reports
 ### November 22, 2022 (DESCW-726) Quarterly Review Report
 - Backend
   - Added controller function, model functions, routes for Project Quarterly Review Report